% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_params.R
\name{expand_params}
\alias{expand_params}
\title{Expand parameters from OUwie output}
\usage{
expand_params(results)
}
\arguments{
\item{results}{output from the ouwie_tidy function}
}
\description{
This function expands the output from ouwie_tidy so that parameter values for all regimes are visible.
}
\examples{
#simulating dummy data to run through OUwie
phy <- pbtree(n = 20, nsim = 2)
disc_trait <- setNames(sample(letters[1:2], 20, replace = TRUE), phy[[1]]$tip.label)
cont_traits <- as_tibble(iris[1:20, 1:2]) \%>\% 
 mutate(species = phy[[1]]$tip.label)

models <- c("BM1", "BMS", "OUM") #set of models to run on each simmap

results <- ouwie_tidy(phy, disc_trait, cont_traits, models, nsim = 2)#' 
expand_params(results)
}
