% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouwie_tidy.R
\name{ouwie_tidy}
\alias{ouwie_tidy}
\title{Wrapper OUwie function}
\usage{
ouwie_tidy(
  phy,
  disc_trait,
  cont_traits,
  models,
  nsim,
  tip_col = NULL,
  dir = NULL,
  ...
)
}
\arguments{
\item{phy}{Phylogenetic tree(s) containing all species/specimens present in dataset.}

\item{disc_trait}{Named vector of discrete characters where the names match the tree tips}

\item{cont_traits}{Data frame containing a character column with tree tips and any number of columns containing continuous trait data. Order of columns does not matter.}

\item{models}{Vector containing the set of models to run on each tree as designated in OUwie documentation}

\item{nsim}{Number of simmaps to create and run all analyses on}

\item{tip_col}{Optional; column name in "cont_traits" that matches tree tips}

\item{dir}{Optional; the directory to save output to, otherwise will save to working directory}

\item{params}{Optional; list of parameters to be passed to OUwie. NOTE: be aware that default OUwie parameters differ from the default OUwie documentation.}
}
\value{
Function returns a list of the following
\item{input}{Simmaps and traits originally input into OUwie}
\item{full_output}{Raw tibble with all results}
\item{tidy_output}{Tidy tibble summarizing results}
}
\description{
This function allows you to run OUwie across all combinations of trees, simmaps, models, and continuous traits and outputs analyses in a neat tibble.
}
\examples{
#simulating dummy data to run through OUwie
phy <- pbtree(n = 20, nsim = 2)
disc_trait <- setNames(sample(letters[1:2], 20, replace = TRUE), phy[[1]]$tip.label)
cont_traits <- as_tibble(iris[1:20, 1:2]) \%>\%
 mutate(species = phy[[1]]$tip.label)

models <- c("BM1", "OUM") #set of models to run on each simmap

results <- ouwie_tidy(phy, disc_trait, cont_traits, models, nsim = 2)
results$tidy_output
}
\references{
Beaulieu, J. M., & O’Meara, B. 2014. OUwie: analysis of evolutionary rates in an OU framework.
}
