% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_params.R
\name{plot_params}
\alias{plot_params}
\title{Plot parameters from OUwie output}
\usage{
plot_params(results)
}
\arguments{
\item{results}{output from the ouwie_tidy function}
}
\value{
Function returns a plot summarizing the model comparisons for each continuous trait in the results output
}
\description{
This function plots the output from ouwie_tidy so that all regimes are visible. Separate plots created for each continuous trait
}
\examples{
#simulating dummy data to run through OUwie
phy <- pbtree(n = 20, nsim = 2)
disc_trait <- setNames(sample(letters[1:2], 20, replace = TRUE), phy[[1]]$tip.label)
cont_traits <- as_tibble(iris[1:20, 1:2]) \%>\%
 mutate(species = phy[[1]]$tip.label)

models <- c("BM1", "OUM") #set of models to run on each simmap

results <- ouwie_tidy(phy, disc_trait, cont_traits, models, nsim = 2)#'
plot_params(results)
}
\references{
Beaulieu, J. M., & O’Meara, B. 2014. OUwie: analysis of evolutionary rates in an OU framework.
}
